requirements/requirements_base.txt
# ========================================
# DEPENDENCIAS BASE - PERFOROSVEN
# ========================================
# Estas dependencias se usan en desarrollo y producción

# Django Framework
Django==5.2.7

# Base de datos PostgreSQL
psycopg2-binary==2.9.9

# Django REST Framework para APIs
djangorestframework==3.15.2

# Gestión de variables de entorno
django-environ==0.11.2

# Manejo de archivos Excel
openpyxl==3.1.5
pandas==2.2.3

# Procesamiento de imágenes
Pillow==11.0.0

# Utilidades de fecha/hora
python-dateutil==2.9.0

# Validación de datos
django-crispy-forms==2.3
crispy-tailwind==1.0.3
